+++ Emulator/RetroShell/RetroShell.cpp
@@ -454,7 +454,7 @@ RetroShell::dump(C64Component &component, Category category)
 void
 RetroShell::vsyncHandler()
 {
-    if (cpu.clock >= (u64)wakeUp) {
+    if (cpu.clock >= wakeUp) {
         
         msgQueue.put(MSG_SCRIPT_WAKEUP);
         wakeUp = INT64_MAX;
